Moved SeoInterfaces in parent directory / namespace to share with other generators
