+++ Seo/Basic/BasicSeoInterface.php
@@ -2,6 +2,10 @@
 
 namespace Leogout\Bundle\SeoBundle\Seo\Basic;
 
+use Leogout\Bundle\SeoBundle\Seo\TitleSeoInterface;
+use Leogout\Bundle\SeoBundle\Seo\DescriptionSeoInterface;
+use Leogout\Bundle\SeoBundle\Seo\KeywordsSeoInterface;
+
 /**
  * Description of BasicSeoInterface.
  *
